requirements.txt
# Combined requirements for local development
# Core ML/DL
torch==2.1.0
torchvision==0.16.0
ultralytics>=8.3.0

# Data processing
numpy==1.24.3
pandas==2.0.3
opencv-python==4.8.0.76
Pillow>=7.1.0,<10
albumentations==1.3.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1
streamlit==1.25.0

# Training utilities
tensorboard==2.14.0
tqdm==4.66.1
pyyaml==6.0.1

# Utilities
scipy==1.11.2
scikit-learn==1.3.0
